refactor: make the program importable for a downstream dev CLI
Split building the Commander program from parsing argv so a private dev
CLI can `import { program } from '@withone/cli'`, register its own
commands, and parse — without this package running on import.

- src/index.ts -> src/cli.ts: builds and exports `program`, no longer
  calls `.parse()`; dropped the now-irrelevant shebang
- new src/index.ts: thin package entry that re-exports `program`
- bin/cli.js: imports `program` and calls `.parse()` — the only place
  argv is parsed
- package.json: add `exports` map (types + import -> dist/index.*) so
  the package is importable; bump version 1.43.4 -> 1.43.5

No behavior change: verified build, --version, --help, nested
subcommand dispatch, --agent JSON mode, network GETs (whoami,
platforms), and the unknown-command error path.
