+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@withone/cli",
-  "version": "1.43.4",
+  "version": "1.43.5",
   "description": "CLI for managing One",
   "type": "module",
   "files": [
@@ -11,6 +11,12 @@
   ],
   "bin": {
     "one": "./bin/cli.js"
+  },
+  "exports": {
+    ".": {
+      "types": "./dist/index.d.ts",
+      "import": "./dist/index.js"
+    }
   },
   "scripts": {
     "build": "tsup",
